+++ tests/Storage/StorageClientTest.php
@@ -272,10 +272,10 @@ class StorageClientTest extends TestCase
         $storeName = $this->TEST_STORE_NAME;
         $key = uniqid();
         $value = 42;
-        $response = $this->client->putInteger($storeName, $key, $value);
+        $response = $this->client->putInt($storeName, $key, $value);
         $this->assertNull($response->asError());
         $this->assertNotNull($response->asSuccess(), "Expected a success but got: $response");
-        $this->setGetDeleteTestCommon($storeName, $key, $value, StorageValueType::INTEGER);
+        $this->setGetDeleteTestCommon($storeName, $key, $value, StorageValueType::INT);
     }
 
     public function testSetGetDeleteDouble()
@@ -283,10 +283,10 @@ class StorageClientTest extends TestCase
         $storeName = $this->TEST_STORE_NAME;
         $key = uniqid();
         $value = 3.14;
-        $response = $this->client->putDouble($storeName, $key, $value);
+        $response = $this->client->putFloat($storeName, $key, $value);
         $this->assertNull($response->asError());
         $this->assertNotNull($response->asSuccess(), "Expected a success but got: $response");
-        $this->setGetDeleteTestCommon($storeName, $key, $value, StorageValueType::DOUBLE);
+        $this->setGetDeleteTestCommon($storeName, $key, $value, StorageValueType::FLOAT);
     }
 
     private function setGetDeleteTestCommon($storeName, $key, $value, string $type)
@@ -305,13 +305,13 @@ class StorageClientTest extends TestCase
             $this->assertEquals($response->valueString(), $value);
         }
 
-        if ($type != StorageValueType::INTEGER) {
+        if ($type != StorageValueType::INT) {
             $this->assertEquals(null, $response->valueInteger());
         } else {
             $this->assertEquals($response->valueInteger(), $value);
         }
 
-        if ($type != StorageValueType::DOUBLE) {
+        if ($type != StorageValueType::FLOAT) {
             $this->assertEquals(null, $response->valueDouble());
         } else {
             $this->assertEquals($response->valueDouble(), $value);
